app/requirements.txt
django==3.2
Pillow
django-admin-tailwind
django-paypal
python-decouple